requirements.txt
Flask==2.1.2
beautifulsoup4==4.11.1
pandas==1.4.2
python-dotenv==0.19.2
supabase==2.5.1
